veeam: add hierarchy_name property to job resource
We may have multiple hierarchy roots in Veeam (multiple VMware hosts, etc).
If the VM name is not found in the first hierarchy root, Terraform will
throw an error.
Add a vm_hierarchy_name property to the veeam_job resource in order to
configure the hierarchy root to lookup when adding a VM to a job.

Signed-off-by: Fatih Acar <fatih.acar@nerim.com>